docs: fix dtype issue in doc pages

+++ docs/tutorial.py
@@ -131,14 +131,14 @@ pprint(ome_zarr_image)
 #
 # For numpy arrays:
 
-arr0 = np.zeros(shape=(100, 100), dtype=np.uint16)
-arr1 = np.zeros(shape=(50, 50), dtype=np.uint16)
+arr0 = np.zeros(shape=(100, 100), dtype=np.uint8)
+arr1 = np.zeros(shape=(50, 50), dtype=np.uint8)
 array_specs = [ArraySpec.from_array(arr0), ArraySpec.from_array(arr1)]
 
 # or for Zarr arrays:
 
-arr_zarr0 = zarr.zeros(shape=(100, 100), dtype=np.uint16, zarr_format=2)
-arr_zarr1 = zarr.zeros(shape=(50, 50), dtype=np.uint16, zarr_format=2)
+arr_zarr0 = zarr.zeros(shape=(100, 100), dtype=np.uint8, zarr_format=2)
+arr_zarr1 = zarr.zeros(shape=(50, 50), dtype=np.uint8, zarr_format=2)
 array_specs = [ArraySpec.from_array(arr_zarr0), ArraySpec.from_array(arr_zarr1)]
 
 # ## Saving datasets
